Move parenthesis removal tests to printer subdirectory
