+++ test/unit/printer/deadCodeAfterReturn.spec.ts
